+++ lib/jekyll/tags/include.rb
@@ -129,6 +129,9 @@ eos
 =======
       def render(context)
         dir = File.join(File.realpath(context.registers[:site].source), INCLUDES_DIR)
+<<<<<<< HEAD
+>>>>>>> origin/v1-stable
+=======
 >>>>>>> origin/v1-stable
 
       def render(context)
@@ -146,6 +149,7 @@ eos
         path = File.join(dir, file)
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
         validate_path(path, dir, site.safe)
 
         # Add include to dependency tree
@@ -158,6 +162,9 @@ eos
 =======
         validate_path(path, dir, context.registers[:site].safe)
 >>>>>>> jekyll/v1-stable
+=======
+        validate_path(path, dir, context.registers[:site].safe)
+>>>>>>> origin/v1-stable
 =======
         validate_path(path, dir, context.registers[:site].safe)
 >>>>>>> origin/v1-stable
@@ -174,6 +181,7 @@ eos
         end
       end
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
       def load_cached_partial(path, context)
@@ -225,6 +233,18 @@ eos
 
       def path_relative_to_source(dir, path)
         File.join(@includes_dir, path.sub(Regexp.new("^#{dir}"), ""))
+=======
+      def validate_path(path, dir, safe)
+        if safe && !realpath_prefixed_with?(path, dir)
+          raise IOError.new "The included file '#{path}' should exist and should not be a symlink"
+        elsif !File.exist?(path)
+          raise IOError.new "Included file '#{path}' not found"
+        end
+      end
+
+      def realpath_prefixed_with?(path, dir)
+        File.exist?(path) && File.realpath(path).start_with?(dir)
+>>>>>>> origin/v1-stable
       end
 
       def realpath_prefixed_with?(path, dir)