+++ test/test_generated_site.rb
@@ -23,6 +23,7 @@ class TestGeneratedSite < JekyllUnitTest
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
       assert_equal 48, @site.posts.size
 =======
       assert_equal 37, @site.posts.size
@@ -57,6 +58,9 @@ class TestGeneratedSite < JekyllUnitTest
 =======
       assert_equal 49, @site.posts.size
 >>>>>>> jekyll/master
+=======
+      assert_equal 37, @site.posts.size
+>>>>>>> origin/v1-stable
     end
 
     should "insert site.posts into the index" do