+++ lib/jekyll/page.rb
@@ -140,9 +140,18 @@ module Jekyll
     #
     # Returns the destination file path String.
     def destination(dest)
+<<<<<<< HEAD
       path = site.in_dest_dir(dest, URL.unescape_path(url))
       path = File.join(path, "index") if url.end_with?("/")
+<<<<<<< HEAD
       path << output_ext unless path.end_with? output_ext
+=======
+      path <<  output_ext unless path.end_with?(output_ext)
+=======
+      path = Jekyll.sanitized_path(dest, url)
+      path = File.join(path, "index.html") if url =~ /\/$/
+>>>>>>> jekyll/v1-stable
+>>>>>>> pod/jekyll-gemfile
       path
     end