+++ test/test_generated_site.rb
@@ -12,7 +12,15 @@ class TestGeneratedSite < JekyllUnitTest
     end
 
     should "ensure post count is as expected" do
+<<<<<<< HEAD
       assert_equal 49, @site.posts.size
+=======
+<<<<<<< HEAD
+      assert_equal 48, @site.posts.size
+=======
+      assert_equal 37, @site.posts.size
+>>>>>>> jekyll/v1-stable
+>>>>>>> pod/jekyll-gemfile
     end
 
     should "insert site.posts into the index" do