+++ History.markdown
@@ -941,6 +941,7 @@
 
 ### Site Enhancements
 
+<<<<<<< HEAD
   * Document Kramdown's GFM parser option (#1791)
   * Move CSS to includes & update normalize.css to v2.1.3 (#1787)
   * Minify CSS only in production (#1803)
@@ -1020,15 +1021,21 @@
 
 =======
 >>>>>>> jekyll/v1-stable
+=======
+>>>>>>> origin/v1-stable
 ## 1.5.1 / 2014-03-27
 
 ### Bug Fixes
 
 <<<<<<< HEAD
+<<<<<<< HEAD
   * Only strip the drive name if it begins the string (#2176)
 =======
 * Only strip the drive name if it begins the string (#2176)
 >>>>>>> jekyll/v1-stable
+=======
+* Only strip the drive name if it begins the string (#2176)
+>>>>>>> origin/v1-stable
 
 ## 1.5.0 / 2014-03-24
 
@@ -1045,10 +1052,14 @@
 ### Development Fixes
 
 <<<<<<< HEAD
+<<<<<<< HEAD
   * Lock `cucumber` at `1.3.11` (#2167)
 =======
   * Lock `cucmber` at `1.3.11` (#2167)
 >>>>>>> jekyll/v1-stable
+=======
+  * Lock `cucmber` at `1.3.11` (#2167)
+>>>>>>> origin/v1-stable
 
 ## 1.4.3 / 2014-01-13
 
@@ -1057,6 +1068,7 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 
 >>>>>>> jekyll/v1-stable
@@ -1069,17 +1081,24 @@
 =======
 
 >>>>>>> jekyll/master
+=======
+
+>>>>>>> origin/v1-stable
   * Patch show-stopping security vulnerabilities (#1944)
 
 ## 1.4.2 / 2013-12-16
 
 ### Bug Fixes
+<<<<<<< HEAD
 
+=======
+>>>>>>> origin/v1-stable
   * Turn on Maruku fenced code blocks by default (#1830)
 
 ## 1.4.1 / 2013-12-09
 
 ### Bug Fixes
+<<<<<<< HEAD
 
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -1142,6 +1161,8 @@
 >>>>>>> jekyll/master
 =======
 >>>>>>> jekyll/master
+=======
+>>>>>>> origin/v1-stable
   * Don't allow nil entries when loading posts (#1796)
 
 ## 1.4.0 / 2013-12-07