+++ History.markdown
@@ -60,6 +60,10 @@
 =======
   * Move Cucumber to using RSpec-Expections and furthering JRuby support. (#4343)
 <<<<<<< HEAD
+<<<<<<< HEAD
+>>>>>>> jekyll/master
+=======
+  * Rearrange Cucumber and add some flair. (#4347)
 >>>>>>> jekyll/master
 =======
   * Rearrange Cucumber and add some flair. (#4347)